skills/f5-allinone/requirements.txt
requests>=2.28.0
urllib3>=1.26.0
pyyaml>=6.0
pytest>=7.0.0
pytest-mock>=3.10.0
